Dump renderer log to working directory on error
When the renderer crashes, the log usually tells you what's wrong.
However, you can't see that information without re-running the client
with `--verbose`. If the issue is intermittent, or if you receive
several long jobs in a row, it may take many hours before you find out
what the issue is.

This change keeps the log in memory and tries to save it to disk in the
event of an error, to make it easier to diagnose issues. While this does
increase memory usage somewhat, my testing shows that a typical job's
output doesn't amount to a significant amount of data, and I expect the
StringBuilder's initial 64KB capacity should be sufficient for jobs that
run for less than 5 minutes or so.