Feat: update progress bar when application is minimised on the system tray
Include a call to the setImageAutoSize() method to ensure proper refresh in all versions of Windows as well as MacOS (due to OSX menubar handle and lack of proper support from Mac Java version the icon might not be updated on several version os MacOS). The proper way to support MacOS is to compile the client for Java 9+ versions.