Fix: revert to base icon when rendering process is interrupted with an exception
Fixed the problem where when the rendering progress is interrupted in the middle of the execution (currently because the memory required by the process is larger than the available one OR there is a problem processing the log ingestion) the application icon is not reset to the base one. Now icon is reset to base if any error happens in the middle of the rendering progress.