fix: background uploads queuing not working properly
Fix the sync/async detection to ensure that handle the server side projects properly. The routine was incorrectly uploading sync files with the async routine and vice-versa.
Also, in agreement with Laurent, removed the _-max-uploading-job_ inline option to make the client sending the completed frames in background all the time.