Fix: Solve the Blender CLI threading issue
Currently, the -t parameter in Blender CLI indicates Blender how many threads (1-1024 or 0 for all the available processor cores) should use for rendering and other operations. However, the -t parameter is overwritten by the configuration of the scene .blend file.

This PR sets by code the number of cores as specified in the configuration of the App.